move plan config to utils
